Improve migration flow
